next attempt of a translator for #10
The original approach was too intensive  It started with the list of
words and tried a straight regex find replace.  This will work but
extracting the list of spells and cleaning it up will be annoying.

The new approach to use GSTranslateString instead iterates over the
words of the macro instead.